+++ app/Filament/Resources/EmailTemplateResource.php
@@ -25,7 +25,7 @@ class EmailTemplateResource extends Resource
 
     protected static ?string $navigationIcon = 'heroicon-o-envelope';
 
-    protected static ?string $navigationGroup = 'System';
+    protected static ?string $navigationGroup = 'System Settings';
 
     protected static ?int $navigationSort = 40;